refactor(front): rename relative date unit constants to match exports
The constant files exported value arrays (RELATIVE_DATE_UNITS /
RELATIVE_DATETIME_UNITS) but were still named "...SelectOptions.ts" from
when they held select-option objects. Rename them to match their exports.
