+++ packages/twenty-front/src/modules/ui/input/components/internal/date/constants/RelativeDateTimeUnits.ts
@@ -1,4 +1,4 @@
-import { RELATIVE_DATE_UNITS } from '@/ui/input/components/internal/date/constants/RelativeDateUnitSelectOptions';
+import { RELATIVE_DATE_UNITS } from '@/ui/input/components/internal/date/constants/RelativeDateUnits';
 import { type RelativeDateFilterUnit } from 'twenty-shared/utils';
 
 export const RELATIVE_DATETIME_UNITS: RelativeDateFilterUnit[] = [
